Exception handling

+++ src/main/java/com/nato/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.nato.dscommerce.services;
 import com.nato.dscommerce.repositories.ProductRepository;
 import com.nato.dscommerce.dto.ProductDTO;
 import com.nato.dscommerce.entities.Product;
+import com.nato.dscommerce.services.exceptions.DatabaseException;
+import com.nato.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -19,8 +24,9 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Optional<Product> product = repository.findById(id);
-        ProductDTO dto = new ProductDTO(product.get());
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Product not found"));
+        ProductDTO dto = new ProductDTO(product);
         return dto;
     }
 
@@ -40,15 +46,27 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product product = repository.getReferenceById(id);
-        updateData(product, dto);
-        product = repository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = repository.getReferenceById(id);
+            updateData(product, dto);
+            product = repository.save(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Id not found " + id);
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void updateData(Product product, ProductDTO dto) {
